Rewrite to use setup-toml-test
This rewrites the lot to use setup-toml-test[1], rather than relying on
the correct environment being set up on the local machine.

This way the authors of implementations deal with the correct setup of
the ecosystem and building of the required tools. Managing a few
ecosystems is okay, but it's becoming unwieldy to manage over a dozen.
Some ecosystems are fairly stable, others a lot more fickle.

The main downside is that we lose the benchmarks. I might look into that
in the future, but I'm okay with that for now.

Another question is what to do with unmaintained implementations. I
think I'll probably add a feature to take the tests from a fork.

[1]: https://github.com/toml-lang/setup-toml-test